Updated SP framework to be aligned with the September preview. - Updated CashDispenser SP to migrate Storage interface. All cash units related operations. i.e. Start/EndExchange, SetCashUnitInfo commands, CashUnitChanged, CountChanged events are in Storage interface. - Updated CashManagement interface for relocating some of commands from CashDispenser. Those commands can be shared with the CashAcceptor interface. i.e. Controlling shutter, retract items, reporting ECB P6 counterfeit item information, etc. - Updated CardReader SP to migrate Storage interface for the capture card BIN. The updated CardReader SP can be used for CardDispenser as well. - Updated TextTerminal to not controlling LEDs and DisplayLights and control them by the Lights interface. - Supported command timeout handling. - Updated KeyManagement capabilities changes for the TR31 version ID. - Updated E2E Security firmware code. Support for parsing dispense tokens in firmware.

+++ Framework/Core/CashDispenser/Events/DelayedDispenseEvent_g.cs
@@ -29,17 +29,18 @@ namespace XFS4IoT.CashDispenser.Events
         public sealed class PayloadData : MessagePayloadBase
         {
 
-            public PayloadData(int? Delay = null)
+            public PayloadData(double? Delay = null)
                 : base()
             {
                 this.Delay = Delay;
             }
 
             /// <summary>
-            /// The time in milliseconds by which the dispense operation will be delayed.
+            /// The time in seconds by which the dispense operation will be delayed.
+            /// <example>0.1</example>
             /// </summary>
             [DataMember(Name = "delay")]
-            public int? Delay { get; init; }
+            public double? Delay { get; init; }
 
         }
